agent-judge-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
SQLAlchemy==2.0.35
openai==1.51.0
spacy==3.7.5

